Add new pull request generator
