+++ scripts/create-pull-request/package.json
@@ -9,6 +9,7 @@
   "type": "module",
   "author": "pixiv",
   "dependencies": {
-    "@octokit/rest": "^19.0.5"
+    "@octokit/rest": "^20.0.1",
+    "yargs": "^17.7.2"
   }
 }
